fix Plugin item
After some pondering, there are following use-cases:
- standalone dynamic plugin
- standalone static plugin

On macOS, there are 2 more use-cases:
- bundled (dynamic) loadable module
- bundled static framework (aka regular static framework)

Change the Plugin item so it behaves the same as the old
LoadableModule item on macOS except it uses correct plugin
folders for installing.

On macOS, Plugin items are now installed in
/Library/<projectname>/PlugIns. Typically, the path is
/Library/<vendor>/<projectname> but omit vendor for now.

The old LoadableModule item installs binaries to /Library/Frameworks
or /usr/local/lib for compatibility.

Change-Id: I98c246739d3157e02da77b885520f3db3bd6510d